version for teot without order

+++ DiplomaApp.Repositories/Interfaces/IRepository.cs
@@ -16,10 +16,10 @@ namespace DiplomaApp.Repositories.Interfaces
 
         void Create(T item);
 
-        T Read(ushort id);
+        T Read(int id);
 
         void Update(T item);
 
-        void Delete(ushort id);
+        void Delete(int id);
     }
 }
